backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
Whoosh
python-docx
